xxxx

+++ app/Http/Livewire/Settings/BankList.php
@@ -107,54 +107,35 @@ class BankList extends Component
         $this->ville = $bank->ville;
     }
     public function deleteData(){
-        
-        try{
-            $banques = Bank::where('id',$this->banqueID)->first();
-            $banques->delete();
-                        //   if($banqueDeleted){
-                        //     session()->flash('message','banque deleted');
-                        //     $this->selectedBankID = [];
-                    
-                        //       // for hidden the model
-                        //       $this->dispatchBrowserEvent('close-model');
-             
-                        //   }else{
-                        //     session()->flash('error',"Une erreur s'est produite. Veuillez réessayer ");
-                        //     $this->dispatchBrowserEvent('close-model');
-                        //   }
-
-        
-        }catch(QueryException $e){
-            session()->flash('error','vous avez selectionner une banque déja utilisé');
-        }
-
-        // if(count($comptes) > 0){
-        //     foreach ($comptes as  $compte) {
-        //         if($compte->bank_id == $banque->id){
-        //             session()->flash('error','this bank is already on used');
+         $comptes = Compte::all();
+         $banque = Bank::where('id',$this->banqueID)->first();
+        if(count($comptes) > 0){
+            foreach ($comptes as  $compte) {
+                if($compte->bank_id == $banque->id){
+                    session()->flash('error','this bank is already on used');
             
-        //             // for hidden the model
-        //             $this->dispatchBrowserEvent('close-model');
+                    // for hidden the model
+                    $this->dispatchBrowserEvent('close-model');
     
                    
             
-        //         }else{
-        //           $banqueDeleted =  $banque->delete();
-        //           if($banqueDeleted){
-        //             session()->flash('message','banque deleted');
+                }else{
+                  $banqueDeleted =  $banque->delete();
+                  if($banqueDeleted){
+                    session()->flash('message','banque deleted');
             
-        //               // for hidden the model
-        //               $this->dispatchBrowserEvent('close-model');
+                      // for hidden the model
+                      $this->dispatchBrowserEvent('close-model');
      
-        //           }else{
-        //             session()->flash('error',"Une erreur s'est produite. Veuillez réessayer ");
-        //             $this->dispatchBrowserEvent('close-model');
-        //           }
-        //         }
-        //     }
+                  }else{
+                    session()->flash('error',"Une erreur s'est produite. Veuillez réessayer ");
+                    $this->dispatchBrowserEvent('close-model');
+                  }
+                }
+            }
     
          
-        // }
+        }
        
     }
 
@@ -162,58 +143,41 @@ class BankList extends Component
      // DELETE SELECTED BANK START
      public function deleteSelected(){
         $banques = Bank::whereIn('id',$this->selectedBankID)->get();
-        try{
-            $banqueDeleted =  $banques->delete();
-                          if($banqueDeleted){
-                            session()->flash('message','banque deleted');
-                            $this->selectedBankID = [];
-                    
-                              // for hidden the model
-                              $this->dispatchBrowserEvent('close-model');
-             
-                          }else{
-                            session()->flash('error',"Une erreur s'est produite. Veuillez réessayer ");
-                            $this->dispatchBrowserEvent('close-model');
-                          }
+        $comptes = Compte::all();
 
-        
-        }catch(QueryException $e){
-            session()->flash('error','vous avez selectionner une banque déja utilisé');
-        }
-
-        // if($banques && count($comptes) > 0){
+        if($banques && count($comptes) > 0){
             
-        //         foreach ($comptes as  $compte) {
-        //             if( in_array($compte->bank_id,$this->selectedBankID)){
-        //                 session()->flash('error','vous avez selectionner une banque déja utilisé');
+                foreach ($comptes as  $compte) {
+                    if( in_array($compte->bank_id,$this->selectedBankID)){
+                        session()->flash('error','vous avez selectionner une banque déja utilisé');
                 
-        //                 // for hidden the model
-        //                 $this->dispatchBrowserEvent('close-model');
+                        // for hidden the model
+                        $this->dispatchBrowserEvent('close-model');
 
-        //                 $this->selectedBankID = [];
+                        $this->selectedBankID = [];
         
                        
                 
-        //             }else{
-        //               $banqueDeleted =  $banques->delete();
-        //               if($banqueDeleted){
-        //                 session()->flash('message','banque deleted');
-        //                 $this->selectedBankID = [];
+                    }else{
+                      $banqueDeleted =  $banques->delete();
+                      if($banqueDeleted){
+                        session()->flash('message','banque deleted');
+                        $this->selectedBankID = [];
                 
-        //                   // for hidden the model
-        //                   $this->dispatchBrowserEvent('close-model');
+                          // for hidden the model
+                          $this->dispatchBrowserEvent('close-model');
          
-        //               }else{
-        //                 session()->flash('error',"Une erreur s'est produite. Veuillez réessayer ");
-        //                 $this->dispatchBrowserEvent('close-model');
-        //               }
-        //             }
-        //         }
+                      }else{
+                        session()->flash('error',"Une erreur s'est produite. Veuillez réessayer ");
+                        $this->dispatchBrowserEvent('close-model');
+                      }
+                    }
+                }
            
             
     
          
-        // }
+        }
